docs/find-least-similar/requirements.txt
cassandra-driver>=3.25.0,<4
numba
jsonlines
